# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/audit/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Lister les logs applicatifs (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service source",
                        "name": "service",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Trace ID",
                        "name": "traceId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "requestId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date début (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date fin (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ops-server_internal_interfaces_http_response.PaginatedData-models_LogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/audit/trails": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Lister les audit trails (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ressource (ex: user)",
                        "name": "resource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "UUID utilisateur",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date début (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date fin (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ops-server_internal_interfaces_http_response.PaginatedData-models_AuditTrailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audit/trails/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Détail d'un audit trail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID audit trail",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.AuditTrailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AuditTrailResponse": {
            "description": "Représentation publique d'un audit trail",
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "auditId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "newValues": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "oldValues": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "outcome": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "resourceId": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "models.LogResponse": {
            "description": "Représentation publique d'un log applicatif persisté",
            "type": "object",
            "properties": {
                "callerFile": {
                    "type": "string"
                },
                "callerLine": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "level": {
                    "type": "string"
                },
                "logId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "ops-server_internal_interfaces_http_response.PaginatedData-models_AuditTrailResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AuditTrailResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "ops-server_internal_interfaces_http_response.PaginatedData-models_LogResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LogResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}